feat(auth): 서버/클라이언트 컴포넌트 분리 및 인증 미들웨어, protected/layout 적용
