+++ src/app/(protected)/layout.tsx
@@ -1,0 +1,17 @@
+import { cookies } from 'next/headers'
+import { redirect } from 'next/navigation'
+
+export default async function ProtectedLayout({
+  children,
+}: {
+  children: React.ReactNode
+}) {
+  const cookieStore = await cookies()
+  const session = cookieStore.get('SESSION')
+
+  if (!session) {
+    redirect('/login')
+  }
+
+  return <>{children}</>
+}
