+++ src/app/(protected)/mypage/page.tsx
@@ -1,144 +1,5 @@
-'use client'
+import ClientMyPage from './ClientMypage'
 
-import Link from 'next/link'
-import { PATH } from '#app/routes'
-import { Settings, ToggleLeft, ToggleRight } from 'lucide-react'
-import CardTitle from '#components/_common/CardTitle'
-import { Button } from '#components/_common/Button'
-import { useState } from 'react'
-import { useAuthStore } from '#stores/auth/useAuthStore'
-import useTheme from '#contexts/ThemeProvider/hook'
-import { useToggleAlarmMutation } from '#apis/notiClient'
-import { useMyProfileQuery, useDeleteUser } from '#apis/userClient'
-import { ConfirmDialog } from '#components/ui/ConfirmDialog'
-import { useHandleLogout } from '#hooks/useHandleLogout'
-
-export default function MyPage() {
-  const { isLoading: isProfileLoading, isError: isProfileError } =
-    useMyProfileQuery()
-  const { mutate: deleteAccount, isPending: isDeletingAccount } =
-    useDeleteUser()
-  const { user } = useAuthStore()
-  const isLoggedIn = useAuthStore((state) => state.isLoggedIn)
-  const { theme, setTheme } = useTheme()
-
-  const { mutate: toggleAlarm, isPending: isTogglingAlarm } =
-    useToggleAlarmMutation()
-
-  const [showLogoutDialog, setShowLogoutDialog] = useState(false)
-  const [showDeleteDialog, setShowDeleteDialog] = useState<boolean>(false)
-  const handleLogout = useHandleLogout()
-
-  if (!isLoggedIn || isProfileLoading) return null
-
-  return (
-    <CardTitle>
-      <div className="flex flex-col items-center w-full mt-5">
-        <h2 className="text-3xl sm:text-3xl font-bold">마이페이지</h2>
-      </div>
-      <div className="w-full h-[1px] bg-border mt-5"></div>
-
-      <div className="w-full p-5">
-        <div className="flex justify-between items-center mb-4">
-          <p className="text-lg font-semibold">프로필 정보</p>
-          <div>
-            <Link
-              href={PATH.mypageEdit}
-              aria-label="마이페이지 수정 페이지로 이동"
-            >
-              <button>
-                <Settings
-                  className="w-[2rem] h-[2rem] text-gray-500"
-                  aria-label="수정 아이콘"
-                />
-              </button>
-            </Link>
-          </div>
-        </div>
-        <div className="flex justify-between mb-2">
-          <p className="text-gray-600">닉네임</p>
-          <div>{user.nickName}</div>
-        </div>
-        <div className="flex justify-between mb-2">
-          <p className="text-gray-600">이메일</p>
-          <div>{user.email}</div>
-        </div>
-      </div>
-      <hr />
-      <div className="w-full p-5">
-        <div className="flex justify-between items-center mb-4">
-          <p className="text-lg font-semibold">알림 설정</p>
-        </div>
-        <div className="flex justify-between mb-2">
-          <p className="text-gray-600">이메일 결제 리마인드</p>
-          <div>
-            <button onClick={() => toggleAlarm()} disabled={isTogglingAlarm}>
-              {user.alarm ? (
-                <ToggleLeft
-                  aria-label="이메일 결제 리마인드 동의"
-                  className="w-[3rem] h-[3rem] fill-[hsl(var(--primary))] strokeWidth={0} stroke-[hsl(var(--background))]"
-                />
-              ) : (
-                <ToggleRight
-                  aria-label="이메일 결제 리마인드 미동의"
-                  className="w-[3rem] h-[3rem] fill-[hsl(var(--primary))] strokeWidth={0} stroke-[hsl(var(--background))]"
-                />
-              )}
-            </button>
-          </div>
-        </div>
-      </div>
-      <hr />
-      <div className="w-full p-5">
-        <div className="flex justify-between items-center mb-4">
-          <p className="text-lg font-semibold">시스템 설정</p>
-        </div>
-        <div className="flex justify-between mb-2">
-          <p className="text-gray-600">다크모드</p>
-          <button onClick={() => setTheme(theme === 'dark' ? 'light' : 'dark')}>
-            {theme === 'dark' ? (
-              <ToggleRight
-                aria-label="다크모드 버튼 아이콘"
-                className="w-[3rem] h-[3rem] fill-[hsl(var(--primary))] strokeWidth={0} stroke-[hsl(var(--background))]"
-              />
-            ) : (
-              <ToggleLeft
-                aria-label="라이트모드 버튼 아이콘"
-                className="w-[3rem] h-[3rem] fill-[hsl(var(--primary))] strokeWidth={0} stroke-[hsl(var(--background))]"
-              />
-            )}
-          </button>
-        </div>
-
-        <div className="flex justify-end mt-3">
-          <Button
-            onClick={() => setShowLogoutDialog(true)}
-            className="primary hopver:[hsl(var(--primary-hover))]"
-          >
-            <span>로그아웃</span>
-          </Button>
-          <ConfirmDialog
-            open={showLogoutDialog}
-            onOpenChange={setShowLogoutDialog}
-            title="로그아웃 하시겠습니까?"
-            onConfirm={handleLogout}
-          />
-
-          <Button
-            onClick={() => setShowDeleteDialog(true)}
-            className="bg-red-400 hover:bg-red-500 ml-4"
-          >
-            <span>탈퇴하기</span>
-          </Button>
-        </div>
-      </div>
-      <ConfirmDialog
-        open={showDeleteDialog}
-        onOpenChange={setShowDeleteDialog}
-        title="정말 탈퇴하시겠습니까?"
-        description="탈퇴하시면 계정과 기록이 모두 삭제됩니다."
-        onConfirm={() => deleteAccount()}
-      />
-    </CardTitle>
-  )
+export default function MyPagePage() {
+  return <ClientMyPage />
 }
